sin & cos arm realization with fixed point
So. Some time ago I tried to think about sin & cos that I can use on CNC.
This calculation allow to use unsigned uint32_t as argument in range 0..1-1LSB (fixed point)
Or [-0.5 .. 0.5] fixed point. I mean range [0..PI] is scaled into uint32_t [0x00000000..0xFFFFFFFF]
Because 0xFFFFFFFF may be equals to 0.999999999 or equals to -0.999999999 it doesn't metters
I don't know is it true, but think that's it.